# [tests] Speed up test_chunkteacher